refactor: move use cases to dedicated package
