+++ AMW_rest/src/test/java/ch/mobi/itc/mobiliar/rest/tags/usecases/AddTagUseCaseTest.java
@@ -1,4 +1,4 @@
-package ch.mobi.itc.mobiliar.rest.tags;
+package ch.mobi.itc.mobiliar.rest.tags.usecases;
 
 import ch.mobi.itc.mobiliar.rest.dtos.TagDTO;
 import ch.puzzle.itc.mobiliar.business.property.control.PropertyTagEditingService;
